Send sinewave audio
Works when going straight to the PulseAudio output
Not yet gotten it to work with ALSA loopback, not seeing data at the other end
A couple of interruptions per second, possible occational underflow

+++ User/log.py
@@ -57,7 +57,7 @@ def parse_line(line: str) -> dict:
         try:
             e['sequence_no'] = int(e['seq'])
             d = base64.b64decode(e['data'])
-            d = numpy.frombuffer(d, dtype='>i2').astype('int16')
+            d = numpy.frombuffer(d, dtype='<i2').astype('int16')
             e['samples'] = d
             e['event'] = event
